complete table class with dinamic query

+++ Controller/Category.php
@@ -13,7 +13,7 @@ class Controller_Category extends Controller_Core_Action
 	{
 		$modelCetegory = $this->getModelCetegory();
 		$sql = "SELECT * FROM `categories` ORDER BY `categories`.`parent_id` ASC";
-		$categories = $modelCetegory->fetchAll($sql);
+		$categories = $modelCetegory->fetchAll($sql);	
 		$this->setCategory($categories);
 		$this->getTemplete('category/grid.phtml');
 	}
@@ -21,13 +21,18 @@ class Controller_Category extends Controller_Core_Action
 	{
 		$modelCetegory = $this->getModelCetegory();
 		$categories = $modelCetegory->fetchAll();
+		if(!$categories)
+		{
+			throw new Exception("data not found.", 1);
+		}
+
 		$this->setCategory($categories);
 		$this->getTemplete('category/add.phtml');
 	}
 	public function editAction()
 	{
 		$request = $this->getRequest();
-		$id=$request->getParam('category_id');
+		$id=(int)$request->getParam('category_id');
 		if(!isset($id))
 		{
 		throw new Exception("invalid product id.", 1);
@@ -35,14 +40,27 @@ class Controller_Category extends Controller_Core_Action
 
 		$modelCetegory =$this->getModelCetegory();
 		$categories = $modelCetegory->fetchAll();
+		if(!$categories)
+		{
+			throw new Exception("data not found.", 1);
+		}
 		$this->setCategory($categories);
 		$category =$modelCetegory->fetchRow($id);
+		if(!$category)
+		{
+			throw new Exception("data not found.", 1);
+		}
 		$this->setEditCategory($category);
 		$this->getTemplete('category/edit.phtml');
 	}
 	public function insertAction()
 	{
 		$request = $this->getRequest();
+		if (!$request->isPost())
+		{
+			throw new Exception("invalid Request.", 1);
+		}
+
 		$category = $request->getPost('category');
 		print_r($category);	
 		if($category["parent_id"] == 'null')
@@ -52,16 +70,22 @@ class Controller_Category extends Controller_Core_Action
 
 		$modelCetegory = $this->getModelCetegory();
 		$insert=$modelCetegory->insert($category);
-		return $this->redirect("http://localhost/new_project/index.php?a=grid&c=category");
+		return $this->redirect("http://localhost/project-krushal-vataliya/index.php?a=grid&c=category");
 	}
 	public function deleteAction()
 	{
 		$request = $this->getRequest();
-		$id = $request->getParam('category_id');
+		$id =(int)$request->getParam('category_id');
+		if(!$id)
+		{
+			throw new Exception("invalid category ID.", 1);
+		}
+		
 		$modelCetegory = $this->getModelCetegory();
 		$delete = $modelCetegory->delete($id);
-		return $this->redirect("http://localhost/new_project/index.php?a=grid&c=category");
+		return $this->redirect("http://localhost/project-krushal-vataliya/index.php?a=grid&c=category");
 	}
+
 	public function updateAction()		
 	{
 		$request = $this->getRequest();
@@ -74,7 +98,7 @@ class Controller_Category extends Controller_Core_Action
 		}
 
 		$update = $modelCetegory->update($category, $category['category_id']);
-		return $this->redirect("http://localhost/new_project/index.php?a=grid&c=category");
+		return $this->redirect("http://localhost/project-krushal-vataliya/index.php?a=grid&c=category");
 	}
 	
     public function getCategory()
